Moved jira webhook into its own file jira.py
