+++ zproject/urls.py
@@ -156,7 +156,7 @@ urlpatterns += patterns('zerver.views',
     # These are integration-specific web hook callbacks
     url(r'^api/v1/external/beanstalk$',     'webhooks.beanstalk.api_beanstalk_webhook'),
     url(r'^api/v1/external/github$',        'webhooks.github.api_github_landing'),
-    url(r'^api/v1/external/jira$',          'webhooks.api_jira_webhook'),
+    url(r'^api/v1/external/jira$',          'webhooks.jira.api_jira_webhook'),
     url(r'^api/v1/external/pivotal$',       'webhooks.api_pivotal_webhook'),
     url(r'^api/v1/external/newrelic$',      'webhooks.api_newrelic_webhook'),
     url(r'^api/v1/external/bitbucket$',     'webhooks.api_bitbucket_webhook'),
